+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # @koopjs/geoservice-utils
 
+## Unreleased
+### Adds
+- combineObjectIdsAndWhere utility
+
 ## 2.0.0
 ### Major Changes
 - breaking change, rename "normalize" to "standardize"